Remove the failing object identity test.
The test attemts to index Lua tables with Luabind objects and tests if logically equal objects are treated as denoting the same table element when used as key. But that is not and *cannot* be the case, as the Lua Reference Manual points out (http://www.lua.org/manual/5.2/manual.html#2.1):

> The indexing of tables follows the definition of raw equality in the
> language. The expressions a[i] and a[j] denote the same table element if and
> only if i and j are raw equal (that is, equal without metamethods).

Note especially the last, parenthesized statement.

It follows that the test does not show a bug in Luabind but rather the "limitations" of the Lua language.

Credits to @Oberon00