Support registration of unnamed classes.
Unnamed classes are registered with Luabind, but inaccessible from Lua.

This is useful e.g. for registering multiple instantiations of a class template, and construct a matching instance using a factory function, e.g. boost::make_shared.

Note that the class representation, which is stored in the Lua interpreter, is kept alive by a Lua reference stored in the class representation itself.

Credits to Peter Colberg