Fix yielding with yield_policy.
Previsouly the lua_yield call was made on a stack that potentially
contained objects with nontrivial destructors (although probably not in
practice).
Also, since lua_yield is implemented by throwing an exception when Lua
is compiled as C++, the catch(...) in
detail::function_object_impl::entry_point (make_function.hpp) would
catch the yield-exception and convert it into a lua_error call which
blows everything up.

Credits to @Oberon00