Fix overload candidates overflow.
Overload candidates still reside in a statically allocated raw array, but now
overflows are checked. Since the array (except for the first element if it is
the only one) is only used to generate the ambiguity error message, just
discard additional elements, keeping track of how many there were and display
that number in the error message if > 0.

This fixes https://github.com/Oberon00/luabind/issues/25

Credits to @Oberon00